pyproject.toml
[project]
name = "schedule-planner"
version = "0.1.0"
description = "Schedule Planner — backend API and scheduling engine"
readme = "README.md"
requires-python = ">=3.11,<3.13"

dependencies = []

[project.optional-dependencies]
# Core (src/): data processing and AI client
pipeline = [
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "pydantic>=2.0.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "numpy>=2.0.0",
]

# Backend API (backend/): FastAPI, DB, auth
backend = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.9",
    "sqlalchemy>=2.0.0",
    "alembic>=1.14.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "email-validator>=2.0.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "requests>=2.28.0",
    "xai_sdk>=0.1.0",
    "google-genai>=1.0.0",
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.0",
    "pymupdf>=1.24.0",
    "ortools>=9.15.6755",
]
